+++ app/app.py
@@ -1,9 +1,18 @@
 #!/usr/bin/env python
+'''
+
+'''
+from sys import exit
+from modules.loadSaveHF import load_and_save_model, load_and_save_dataset, \
+    load_and_save_tokenizer, load_and_save_metrics
 
 class App():
-    print()
+    '''
+    '''
+    
 
-main = launch_new_instance = App.launch_instance
+# main = launch_new_instance = App.launch_instance
 
 if __name__ == "__main__":
-    main()
+    # main()
+    exit(App.launch_instance)